IT WORKS!!!
I can play tones, and the reciever makes the corresponding LED light up. And if no tone is played, the last lit LED fades to black. ISSUES: Not perfect FFT calculation, often flickering between two LEDs, and it also often think it's the tone C when it's not. But WOW!